Rename `app` package into `teetaskruntime`
