+++ src/itest/java/com/iexec/sms/secret/TeeTaskRuntimeSecretIntegrationTests.java
@@ -20,10 +20,10 @@ import com.iexec.common.utils.HashUtils;
 import com.iexec.sms.ApiClient;
 import com.iexec.sms.CommonTestSetup;
 import com.iexec.sms.encryption.EncryptionService;
-import com.iexec.sms.secret.app.OnChainObjectType;
-import com.iexec.sms.secret.app.SecretOwnerRole;
-import com.iexec.sms.secret.app.TeeTaskRuntimeSecret;
-import com.iexec.sms.secret.app.TeeTaskRuntimeSecretRepository;
+import com.iexec.sms.secret.teetaskruntime.OnChainObjectType;
+import com.iexec.sms.secret.teetaskruntime.SecretOwnerRole;
+import com.iexec.sms.secret.teetaskruntime.TeeTaskRuntimeSecret;
+import com.iexec.sms.secret.teetaskruntime.TeeTaskRuntimeSecretRepository;
 import feign.FeignException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
