+++ src/main/java/com/iexec/sms/tee/session/palaemon/PalaemonSessionService.java
@@ -22,10 +22,10 @@ import com.iexec.common.tee.TeeEnclaveConfiguration;
 import com.iexec.common.utils.FileHelper;
 import com.iexec.common.utils.IexecEnvUtils;
 import com.iexec.sms.secret.Secret;
-import com.iexec.sms.secret.app.OnChainObjectType;
-import com.iexec.sms.secret.app.SecretOwnerRole;
-import com.iexec.sms.secret.app.TeeTaskRuntimeSecret;
-import com.iexec.sms.secret.app.TeeTaskRuntimeSecretService;
+import com.iexec.sms.secret.teetaskruntime.OnChainObjectType;
+import com.iexec.sms.secret.teetaskruntime.SecretOwnerRole;
+import com.iexec.sms.secret.teetaskruntime.TeeTaskRuntimeSecret;
+import com.iexec.sms.secret.teetaskruntime.TeeTaskRuntimeSecretService;
 import com.iexec.sms.secret.web2.Web2SecretsService;
 import com.iexec.sms.secret.web3.Web3SecretService;
 import com.iexec.sms.tee.challenge.TeeChallenge;
